results ui updated, just to fix sorting
