+++ client/src/components/Results.js
@@ -10,84 +10,129 @@ import { IoIosArrowDropleft } from "react-icons/io";
 import ReactStars from "react-rating-stars-component";
 import nodeServer from "../api/nodeServer";
 
-const Results = ({ results, history }) => {
-  const [displayResults, setDisplayResults] = useState(results);
+const Results = ({
+  onBackClicked,
+  selectedNeeds,
+  selectedSupportTypes,
+  selectedPersonalisations,
+  postcode,
+  charities,
+}) => {
+  const [prioritisedResults, setprioritisedResults] = useState([]);
+  const [allCharities, setAllCharities] = useState([]);
 
-  //   async getRating(charities) {
-  //     const finalCharities =[]
-  //     await Promise.all(charities.map(charity => {
-  //           // console.log(charity.PlaceID);
-  //           if(charity.PlaceID){
-  //             nodeServer.get(`/googleratings/${charity.PlaceID}`).then((res) => {
-  //               const rating = res.data.rating
-  //               // console.log('rating', rating);
-  //               charity.googleRating = rating
-  //               finalCharities.push(charity)
-  //               this.setState({ finalCharities: finalCharities})
-  //             })
-  //           } else {
-  //             // console.log('fired');
-  //             this.setState({ finalCharities: finalCharities})
-  //           }
-  //  }))
-  //   };
+  useEffect(() => {
+    constructCharityObjects();
+  }, []);
 
-  const getRating = async () => {
-    const finalCharities = [];
-    // console.log(charity.PlaceID);
-    await Promise.all(
-      displayResults.map((charity) => {
-        if (charity.PlaceID) {
-          nodeServer.get(`/googleratings/${charity.PlaceID}`).then((res) => {
-            const rating = res.data.rating;
+  useEffect(() => {
+    sortCharities();
+  }, [allCharities]);
 
-            charity.googleRating = rating;
-            console.log("rating", charity);
-            finalCharities.push(charity);
-            //   this.setState({ finalCharities: finalCharities})
-          });
-        } else {
-          console.log("fired");
-        }
-        // return finalCharities
-      })
-    ).then((res) => {
-      console.log(res);
-    });
-    // setDisplayResults(finalCharities)
+  const sortCharities = () => {
+    // Sort by number of needs met
+    // then by support types
+    // then by personalisations
+    // then by postcode
+
+    const prioritisedCharities = allCharities.sort(
+      (a, b) => needsMet(a, b) || supportTypes(a, b) || personalisations(a, b)
+    );
+
+    setprioritisedResults(prioritisedCharities);
   };
 
-  const showResults = displayResults.map((result) => (
-    <div className="results-list-container" key={result.PlaceID}>
+  const needsMet = (a, b) => b.needsMet.length - a.needsMet.length;
+
+  const supportTypes = (a, b) =>
+    b.typesOfSupportOffered.length - a.typesOfSupportOffered.length;
+
+  const personalisations = (a, b) =>
+    b.personalisationsMet.length - a.personalisationsMet.length;
+
+  const constructCharityObjects = () => {
+    // get unique org ids
+    const uniqueOrgIds = [
+      ...new Set(charities.map((charity) => charity.OrgID)),
+    ];
+
+    // for each org generate list of
+    // userOption (needs)
+    // UserOption_Type (types of support)
+    // personalisation
+
+    const charityObjects = uniqueOrgIds.slice(0, 10).map((orgId) => {
+      const charity = charities.find((charity) => charity.OrgID === orgId);
+
+      const servicesFromCharity = charities.filter(
+        (charity) => charity.OrgID == orgId
+      );
+
+      const needsMet = [
+        ...new Set(servicesFromCharity.map((service) => service.UserOption)),
+      ];
+      const typesOfSupportOffered = [
+        ...new Set(
+          servicesFromCharity.map((service) => service.UserOption_Type)
+        ),
+      ];
+      const personalisationsMet = [
+        ...new Set(
+          servicesFromCharity.map((service) => service.Personalisation)
+        ),
+      ]; // might not work
+
+      delete charity.TypeOfSupport;
+      delete charity.UserOption_Type;
+      delete charity.UserOption;
+      delete charity.Personalisation;
+
+      return {
+        ...charity,
+        needsMet,
+        typesOfSupportOffered,
+        personalisationsMet,
+      };
+    });
+
+    setAllCharities(charityObjects);
+  };
+
+  const PrioritisedListOfCharities = prioritisedResults.map((charity) => (
+    <div className="results-list-container" key={charity.PlaceID}>
       <div className="results-title-container">
-        {result.Logo ? (
-          <a href={result.ServiceURL} target="_blank">
-            <img className="results-list-logo" src={result.Logo} />
+        {charity.Logo ? (
+          <a href={charity.ServiceURL} target="_blank">
+            <img className="results-list-logo" src={charity.Logo} />
           </a>
         ) : (
           <div></div>
         )}
 
         <div className="results-list-title-service">
-          {result.ServiceURL ? (
+          {charity.ServiceURL ? (
             <p className="results-list-title">
               {" "}
-              <a href={result.ServiceURL} target="_blank">
-                {result.OrgName}
+              <a href={charity.ServiceURL} target="_blank">
+                {charity.OrgName}
               </a>
             </p>
           ) : (
-            <p className="results-list-title">{result.OrgName}</p>
+            <p className="results-list-title">{charity.OrgName}</p>
           )}
 
           <p className="results-list-service-description">
-            {result.ServiceDescription}
+            {charity.ServiceDescription}
           </p>
-          {result.googleRating ? (
+          {charity.googleRating ? (
             <p>
               {" "}
-              <ReactStars count={5} value={result.googleRating} isHalf={true} />
-              {result.googleRating}{" "}
+              <ReactStars
+                count={5}
+                value={charity.googleRating}
+                isHalf={true}
+              />
+              {charity.googleRating}{" "}
             </p>
           ) : (
             <p>
@@ -100,36 +145,36 @@ const Results = ({ results, history }) => {
       </div>
 
       <div className="results-button-container">
-        {result.PhoneNo && (
+        {charity.PhoneNo && (
           <button className="results-list-button" type="button">
             <IconContext.Provider
               value={{ className: "results-react-button-icon" }}
             >
-              <a href={"tel:" + result.PhoneNo}>
+              <a href={"tel:" + charity.PhoneNo}>
                 <BiPhone /> Call
               </a>
             </IconContext.Provider>
           </button>
         )}
 
-        {result.ServiceURL && (
+        {charity.ServiceURL && (
           <button className="results-list-button" type="button">
             <IconContext.Provider
               value={{ className: "results-react-button-icon" }}
             >
-              <a href={result.ServiceURL} target="_blank">
+              <a href={charity.ServiceURL} target="_blank">
                 <BiChat /> Chat{" "}
               </a>
             </IconContext.Provider>
           </button>
         )}
 
-        {result.EmailAddress && (
+        {charity.EmailAddress && (
           <button className="results-list-button" type="button">
             <IconContext.Provider
               value={{ className: "results-react-button-icon" }}
             >
-              <a class="mailto" href={"mailto:" + result.EmailAddress}>
+              <a class="mailto" href={"mailto:" + charity.EmailAddress}>
                 <BiEnvelope /> Email
               </a>
             </IconContext.Provider>
@@ -137,7 +182,7 @@ const Results = ({ results, history }) => {
         )}
       </div>
 
-      {result.OpeningTime && (
+      {charity.OpeningTime && (
         <p className="results-list-sub-title">
           <IconContext.Provider
             value={{ className: "results-react-icons-detail" }}
@@ -147,9 +192,9 @@ const Results = ({ results, history }) => {
           </IconContext.Provider>
         </p>
       )}
-      <p className="results-list-service-detail">{result.OpeningTime}</p>
+      <p className="results-list-service-detail">{charity.OpeningTime}</p>
 
-      {result.PhysicalAddress && (
+      {charity.PhysicalAddress && (
         <p className="results-list-sub-title">
           <IconContext.Provider
             value={{ className: "results-react-icons-detail" }}
@@ -159,33 +204,37 @@ const Results = ({ results, history }) => {
           </IconContext.Provider>
         </p>
       )}
-      <p className="results-list-service-detail">{result.PhysicalAddress}</p>
+      <p className="results-list-service-detail">{charity.PhysicalAddress}</p>
 
-      {result.UserOption && (
-        <p className="results-list-needs-tag">{result.UserOption}</p>
-      )}
+      {/* TODO: REMOVE INLINE CSS */}
+      <div style={{ display: "flex" }}>
+        {charity.needsMet.map((need) => (
+          <span className="results-list-needs-tag">{need}</span>
+        ))}
+      </div>
 
-      {result.Personalisation && (
-        <p className="results-list-personalisation-tag">
-          {result.Personalisation}
-        </p>
-      )}
+      <div style={{ display: "flex" }}>
+        {charity.personalisationsMet.map((personalisation) => (
+          <span className="results-list-personalisation-tag">
+            {personalisation}
+          </span>
+        ))}
+      </div>
 
-      {result.TypeOfSupport && (
+      {charity.typesOfSupportOffered && (
         <p className="results-list-sub-title">Types of Service:</p>
       )}
-      {result.TypeOfSupport && (
-        <li className="results-list-service-detail">{result.TypeOfSupport}</li>
-      )}
 
-      {result.OrgDescription && <p className="results-list-sub-title">About</p>}
-      <p className="results-list-service-detail">{result.OrgDescription}</p>
+      {charity.typesOfSupportOffered.map((supportType) => {
+        return <li className="results-list-service-detail">{supportType}</li>;
+      })}
+
+      {charity.OrgDescription && (
+        <p className="results-list-sub-title">About</p>
+      )}
+      <p className="results-list-service-detail">{charity.OrgDescription}</p>
     </div>
   ));
-
-  // useEffect(() => {
-  //   getRating()
-  // }, [])
 
   return (
     <div className="results-page-container">
@@ -193,14 +242,14 @@ const Results = ({ results, history }) => {
         <div className="title-description-container">
           <h1 className="question-title">Search results</h1>
         </div>
-        <div className="results-page-display">{showResults}</div>
+        <div className="results-page-display">{PrioritisedListOfCharities}</div>
         <div className="bottom-navigation">
-          <IoIosArrowDropleft
-            className="back-button-results"
-            onClick={() => {
-              history.goBack();
-            }}
-          />
+          {onBackClicked && (
+            <IoIosArrowDropleft
+              className="back-button"
+              onClick={onBackClicked}
+            />
+          )}
         </div>
       </div>
     </div>
